+++ src/storages/dataset.js
@@ -6,7 +6,7 @@ import log from '../utils_log';
 
 /* eslint-disable no-unused-vars,import/named,import/no-duplicates,import/order */
 // @ts-ignore
-import ApifyClient from 'apify-client';
+import { ApifyClient } from 'apify-client';
 // @ts-ignore
 import { ApifyStorageLocal } from '@apify/storage-local';
 import { Configuration } from '../configuration';